requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu113
torch==1.12.0
--extra-index-url https://download.pytorch.org/whl/cu113
torchaudio==0.12.0
--extra-index-url https://download.pytorch.org/whl/cu113
torchvision==0.13.0
librosa>=0.9.2
matplotlib>=3.5.1
tqdm>=4.63.0
zipp>=3.7.0
tensorboard>=2.9.1
pillow>=9.0.1
einops>=0.4.1
wget>=3.2
svn>=1.0.1
git+https://github.com/ml-jku/hopfield-layers; platform_system != 'Windows'
